+++ src/bin/www.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata';
 import * as express from 'express';
 import apolloLoader from '../loaders/apollo';
 import expressLoader, { loadHandleError } from '../loaders/express';
-import dbConnection from '../loaders/dbconnect';
+import dbConnection from '../loaders/typeorm';
 import { logger } from '../logger/winston';
 import { config } from '../config';
